requirements.txt
nfl_data_py
pandas
sqlalchemy
psycopg2-binary
scikit-learn
python-dotenv